requirements.txt
asgiref==3.8.1
bleach==6.1.0
Django==4.2
django-debug-toolbar==4.3.0
django-filter==24.2
django-summernote==0.8.20.0
django-taggit==5.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==24.14.0
Markdown==3.6
pillow==10.3.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
webencodings==0.5.1
